Safely remove multiple class directory in Gradle 4.0+ (#51)
Removes current hack, which is highly project and operating system dependent.
Original solution relied on project only containing a 'scala' plugin. Once
the project adds for JVM langiage plugins, it failed.

This fix takes the Gradle version into account and switches to use classesDirs
when the version is 4.0 or newer. It find all of the output directories and then
returns only the one that has a parent folder called 'scala'. This work on the
assumption that Gatling tests are only written in Scala. If a Gatlign source set relies
on additional Java|Groovy|Kotlin classes in the same source set, this will still fail.
The chances of someone doing this is probably very low.